src/py: added baseline classifier and fixed utils bug.
The utils module in the cssigps python package was recording live audio
using 32bit float values, but all of our training examples are made
using 16bit int values. Right now, our features are not robust enough to
handle these differences.

+++ src/py/cssigps/utils.py
@@ -20,7 +20,7 @@ def record_sample(duration=3,samplerate=11025,channels=1,blocking=False):
     blocking : bool
         True if method should block until it completes. False otherwise.
     """
-    sample=sd.rec(duration * samplerate, samplerate=samplerate, channels=channels)
+    sample=sd.rec(duration * samplerate, samplerate=samplerate, channels=channels, dtype='int16')
     if blocking == True:
         sd.wait()
     return sample
